sdk: fix capture integration tests
Because uploading a capture collects everything in the directory, it was including the addendum on the create which caused the data to be different than the base file that the test was comparing against.

I created a seperate folder which has the addendum to use for the update function and removed the addendum from the directory that create uses.